binary_function to module
