+++ polars/polars-lazy/src/prelude.rs
@@ -20,6 +20,7 @@ pub(crate) use crate::{
         aggregation::{AggQuantileExpr, AggregationExpr},
         alias::AliasExpr,
         apply::ApplyExpr,
+        binary_function::BinaryFunctionExpr,
         cast::CastExpr,
         column::ColumnExpr,
         is_not_null::IsNotNullExpr,
